requirements.txt
nltk
pillow
ipython
